Introduce sub social_security_worksheet_data()
RATIONALE: The Social Security Benefits Worksheet can have 3 possible stopping
points (depending on input data): Lines 7, 9 and 18.  For each stopping
point, the user will want to know *both* the stopping point value (to be
entered on 1040 line 6b) and the values of all lines to be filled in up
through the stopping point.

Until now, social_security_benefits() has only returned the stopping
point value -- meaning that the user would have to hack the code to get
a dump of the values for all lines.  We want to create a pretty-printer
for this worksheet as we have done for the Qualified Dividends
worksheet.  To do that, at each stopping point the calculation "engine"
will have to return an array ref of all the lines, and then user-visible
function will have to return either the stopping point value or all the
values.

IMPLEMENTATION: The guts of what in earlier versions was
social_security_benefits() has been extracted into
_social_security_benefits_engine().  This internal subroutine:

* Takes its arguments from either social_security_benefits() or the
  newly introduced user-visible subroutine
social_security_worksheet_data().

* Returns a hashref with two elements keyed on 'taxable_benefits' and
  'worksheet_data' to its caller.

* The caller then returns either a scalar holding the
  2-decimal-place-formatted amount to be entered onto Form 1040 line 6b
or an array reference holding the list of all data to be entered onto
the Social Security Benefits Worksheet.  (But don't forget to check the
boxes on lines 7 and 9!)

The two test files, t/*/001-social.t, have been extended to exercise
social_security_worksheet_data().  Some of the test data has been
compacted to occupy less vertical depth in the terminal.